post install script modified

+++ post-install.sh
@@ -2,24 +2,32 @@
 
 # Installing all necessery packages
 
-sudo zypper install qutebrowser xviewer discord steam mpv htop neofetch ranger kitty wine lutris picom redshift xarchiver opi 7zip libreoffice xrandr xset xinput sensors pavucontrol playerctl python310-adblock
+read -p "Install packages?(y/n)" pcg
+if [ $pcg = "y" ] 
+then
+	sudo zypper install qutebrowser xviewer discord steam mpv htop neofetch ranger kitty wine lutris picom redshift xarchiver opi 7zip libreoffice xrandr xset xinput sensors pavucontrol playerctl python310-adblock cpu-x dmenu
+fi
 
-opi spotify-easyrpm
-
-spotify-easyrpm --set-channel edge
-
-spotify-easyrpm
-
-opi spotify-adblock
+read -p "Install spotify? (y/n)" spotify
+if [ $spotify = "y" ]
+then
+	opi spotify-easyrpm
+	spotify-easyrpm --set-channel edge
+	spotify-easyrpm
+	opi spotify-adblock
+fi
 
 opi j4-dmenu-desktop
-
-opi cpu-x
 
 opi qt5-webengine-widevine
 
 # Moving config files
 
-cp -r .config ~/
+read -p "Replace config files?(y/n)" rplc
 
-cp -r .fonts ~/ 
+if [ $rplc = "y" ]
+then
+	cp -r .config ~/
+
+	cp -r .fonts ~/ 
+fi
